+++ src/main/java/com/svalero/AmazonAA/service/ProductServiceImpl.java
@@ -1,11 +1,15 @@
 package com.svalero.AmazonAA.service;
 
+import com.svalero.AmazonAA.domain.Order;
 import com.svalero.AmazonAA.domain.Product;
 import com.svalero.AmazonAA.domain.Review;
 import com.svalero.AmazonAA.domain.Stock;
 import com.svalero.AmazonAA.domain.dto.ProductDTO;
 import com.svalero.AmazonAA.exception.ProductNotFoundException;
+import com.svalero.AmazonAA.repository.OrderRepository;
 import com.svalero.AmazonAA.repository.ProductRepository;
+import com.svalero.AmazonAA.repository.ReviewRepository;
+import com.svalero.AmazonAA.repository.StockRepository;
 import org.modelmapper.ModelMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -20,6 +24,16 @@ public class ProductServiceImpl implements ProductService{
 
     @Autowired
     private ProductRepository productRepository;
+
+    @Autowired
+    private OrderRepository orderRepository;
+
+    @Autowired
+    private StockRepository stockRepository;
+
+    @Autowired
+    private ReviewRepository reviewRepository;
+
     @Autowired
     ModelMapper modelMapper;
 
@@ -62,10 +76,18 @@ public class ProductServiceImpl implements ProductService{
     }
 
     @Override
-    public void deteleProduct(long id) throws ProductNotFoundException {
+    public boolean deteleProduct(long id) throws ProductNotFoundException {
         Product product = productRepository.findById(id).orElseThrow(ProductNotFoundException::new);
+        List<Order> order = orderRepository.findByProduct_Id(product.getId());
+        List<Review> review = reviewRepository.findByProductReview(product);
+        List<Stock> stock = stockRepository.findByProductStock_Id(product.getId());
+        if(!order.isEmpty() || !review.isEmpty() || !stock.isEmpty()){
+            logger.error("Product " + product + " no se puede borrar ya que referencia a otras tablas en la base de datos, elimina esas referencias antes.");
+            return false;
+        }
         logger.info("Delete Product: " + product);
         productRepository.delete(product);
+        return true;
     }
 
     @Override
